Move the crosshairs to the OSD foreground
BF4.2 introduced foreground and background planes for the OSD,
with the background being used for non-changing elements.
Unfortunately this means that a foreground element such as the
horizon is in front of any background element including the
crosshairs. The resulting blink-out of the crosshairs when the
horizon touches it is deeply annoying, so this change moves
the crosshairs back to the foreground and makes sure they stay
in front of the horizon line.