Add throttle cut
This change provides a dynamic adjustment to the throttle limit based on
battery voltage, with the intent of allowing more control over how
throttle 'feel' changes as the battery discharges. The adjustment is
expressed as a percentage cut that is applied when the battery is full,
along with two voltages to tune the range over which the cut is scaled.
Changes are made in the CLI; throttle_comp_cut is the maximum amount to
reduced the throttle by, expressed as a percentage. throttle_comp_vmax
specifies the voltage above which the full cut percentage applies.
throttle_comp_vmin is the voltage below which no cut is applied. Between
the min and max values the amount of cut is linearly scaled. The
voltages are specified in units of 1/100V, eg 350 is 3.5V.
This feature works in conjunction with the existing throttle limit, so
if a normal throttle limit of 80% is set and then an additional 10% cut
is specified, then at the beginning of the flight throttle cut will make
a further reduction of 8% (10% of the 80%).
The scaling of the cut value is driven from the filtered average cell
voltage, so the behaviour can be influenced by changing the time
constant of the voltage filter using set vbat_lpf_period. The default
value is 3s, I've found 10s works well with throttle compensation.
Limited testing has only been done on a 1S whoop with vmin = 360, vmax =
400 and cut = 20.